requirements.txt
# requirements.txt
# Core SFX agent
elevenlabs>=0.1.0
pydub>=0.25.1
ffmpeg-python>=0.2.0
ruamel.yaml>=0.17.21
mutagen>=1.45.1
pyloudnorm>=0.1.0
gitingest
jinja2>=3.1.2
click>=8.1.3
python-dotenv>=1.0.0
pytest>=7.4.0

#UI
PyQt5>=5.15.9
